linux-arm64 fix -Wshorten-64-to-32

+++ src/mono/mono/mini/jit-icalls.c
@@ -541,7 +541,7 @@ mono_imul_ovf (gint32 a, gint32 b)
 		return 0;
 	}
 
-	return res;
+	return (gint32) res;
 }
 
 gint32
@@ -556,7 +556,7 @@ mono_imul_ovf_un (guint32 a, guint32 b)
 		return 0;
 	}
 
-	return res;
+	return (gint32)res;
 }
 
 gint32
@@ -571,7 +571,7 @@ mono_imul_ovf_un_oom (guint32 a, guint32 b)
 		return 0;
 	}
 
-	return res;
+	return (gint32)res;
 }
 #endif
